move the renderer-only shared modules into the renderer package
