+++ packages/renderer/components/config-switch-field.tsx
@@ -1,8 +1,8 @@
-import { useConfig, useConfigMutation } from "@meru/shared/renderer/react-query";
 import type { Config } from "@meru/shared/types";
 import { Field, FieldContent, FieldDescription, FieldLabel } from "@meru/ui/components/field";
 import { Switch } from "@meru/ui/components/switch";
 import { useIsLicenseKeyValid } from "@/lib/hooks";
+import { useConfig, useConfigMutation } from "@/lib/react-query";
 import { restartRequiredToast } from "@/lib/toast";
 import { ExperimentalFieldBadge } from "./experimental-field-badge";
 import { LicenseKeyRequiredFieldBadge } from "./license-key-required-field-badge";
