+++ packages/renderer/components/popup-window.tsx
@@ -1,6 +1,6 @@
-import { useThemeStore } from "@meru/shared/renderer/theme";
 import { Toaster } from "@meru/ui/components/sonner";
 import { useHotkeys } from "react-hotkeys-hook";
+import { useThemeStore } from "@/lib/theme";
 
 export function PopupWindow({ children }: { children: React.ReactNode }) {
   const theme = useThemeStore((state) => state.theme);
