+++ packages/renderer/recent-download-history.tsx
@@ -1,7 +1,5 @@
 import { MAX_RECENT_DOWNLOAD_HISTORY_ITEMS } from "@meru/shared/constants";
 import { ipc } from "@meru/shared/renderer/ipc";
-import { renderApp } from "@meru/shared/renderer/react";
-import { useConfig, useConfigMutation } from "@meru/shared/renderer/react-query";
 import { Button } from "@meru/ui/components/button";
 import {
   Empty,
@@ -22,6 +20,8 @@ import { cn } from "@meru/ui/lib/utils";
 import { DownloadIcon, FolderIcon, SquareArrowOutUpRightIcon, XIcon } from "lucide-react";
 import { DateFromNow } from "@/components/date-from-now";
 import { PopupWindow } from "@/components/popup-window";
+import { renderApp } from "@/lib/react";
+import { useConfig, useConfigMutation } from "@/lib/react-query";
 
 function RecentDownloadHistory() {
   const { config } = useConfig();
