+++ packages/renderer/routes/settings/blocker.tsx
@@ -1,8 +1,8 @@
-import { useConfig } from "@meru/shared/renderer/react-query";
 import { FieldGroup, FieldSeparator } from "@meru/ui/components/field";
 import { ConfigSwitchField } from "@/components/config-switch-field";
 import { LicenseKeyRequiredBanner } from "@/components/license-key-required-banner";
 import { Settings, SettingsContent, SettingsHeader, SettingsTitle } from "@/components/settings";
+import { useConfig } from "@/lib/react-query";
 
 export function BlockerSettings() {
   const { config } = useConfig();
