+++ packages/renderer/routes/settings/verification-codes.tsx
@@ -1,9 +1,9 @@
-import { useConfig } from "@meru/shared/renderer/react-query";
 import { FieldGroup } from "@meru/ui/components/field";
 import { ConfigSelectField } from "@/components/config-select-field";
 import { ConfigSwitchField } from "@/components/config-switch-field";
 import { LicenseKeyRequiredBanner } from "@/components/license-key-required-banner";
 import { Settings, SettingsContent, SettingsHeader, SettingsTitle } from "@/components/settings";
+import { useConfig } from "@/lib/react-query";
 
 const verificationCodeConfidenceItems = [
   { value: "high", label: "High" },
