+++ packages/renderer/routes/unified-inbox.tsx
@@ -1,8 +1,6 @@
 import { getGoogleDomainFaviconUrl } from "@meru/shared/google";
 import { ms } from "@meru/shared/ms";
-import { createDateTimeFormatter, dayjs } from "@meru/shared/renderer/date";
 import { ipc } from "@meru/shared/renderer/ipc";
-import { useConfig, useConfigMutation } from "@meru/shared/renderer/react-query";
 import { AccountBadge } from "@meru/ui/components/account-badge";
 import { Avatar, AvatarFallback, AvatarGroup, AvatarImage } from "@meru/ui/components/avatar";
 import { Button } from "@meru/ui/components/button";
@@ -40,7 +38,9 @@ import {
 import { useEffect, useMemo, useRef, useState } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 import { SettingsHeader, SettingsTitle } from "@/components/settings";
+import { createDateTimeFormatter, dayjs } from "@/lib/date";
 import { useUnifiedInbox, type UnifiedInboxMessage } from "@/lib/hooks";
+import { useConfig, useConfigMutation } from "@/lib/react-query";
 
 const columnHelper = createColumnHelper<UnifiedInboxMessage>();
