+++ packages/renderer/workspace-app.tsx
@@ -1,6 +1,4 @@
 import { ipc } from "@meru/shared/renderer/ipc";
-import { renderApp } from "@meru/shared/renderer/react";
-import { useConfig } from "@meru/shared/renderer/react-query";
 import type { SupportedWorkspaceApp } from "@meru/shared/workspace-apps";
 import { AccountBadge } from "@meru/ui/components/account-badge";
 import { FindInPage } from "@meru/ui/components/find-in-page";
@@ -16,6 +14,8 @@ import {
 import { WorkspaceAppIcon } from "@meru/ui/components/workspace-app-icon";
 import { DownloadIcon, EllipsisVerticalIcon } from "lucide-react";
 import { useEffect, useState } from "react";
+import { renderApp } from "@/lib/react";
+import { useConfig } from "@/lib/react-query";
 
 function RecentDownloadHistoryButton() {
   return (
