--- packages/shared/renderer/hooks.ts
@@ -1,30 +1,0 @@
-import { ms } from "@meru/shared/ms";
-import { useEffect, useRef, useState } from "react";
-
-export function useCopied() {
-  const [copied, setCopied] = useState(false);
-
-  const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  useEffect(() => {
-    return () => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current);
-      }
-    };
-  }, []);
-
-  const markCopied = () => {
-    setCopied(true);
-
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-
-    timeoutRef.current = setTimeout(() => {
-      setCopied(false);
-    }, ms("2s"));
-  };
-
-  return { copied, markCopied };
-}
